Rework Sensitive in jenkins::cli::exec
This commit changes jenkins::cli::exec from accepting a potentially
Sensitive command parameter to simply marking all commands as Sensitive.
This is because the helper command (at least in our environment)
contains credentials and so all commands invoked by jenkins::cli::exec
need to be redacted.

This commit ends up reverting some of the logic added in the previous
commit and simplifies the logic somewhat.